+++ src/pipex.c
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/28 06:26:25 by fschuber          #+#    #+#             */
-/*   Updated: 2023/11/28 17:42:38 by fschuber         ###   ########.fr       */
+/*   Updated: 2023/11/29 10:33:35 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 	@returns	A file descriptor to the command output. This point to a pipe if
 				out_fd is -1, otherwise it will simply be out_fd.
 */
-static int	execute_command(int in_fd, char *cmd, int out_fd, char **envp)
+int	execute_command(int in_fd, char *cmd, int out_fd, char **envp)
 {
 	int		pipefd[2];
 	pid_t	pid;
@@ -33,6 +33,7 @@ static int	execute_command(int in_fd, char *cmd, int out_fd, char **envp)
 		return (parent_routine(pid, pipefd, out_fd));
 	else
 		exit_error("ERROR: Forking failed.\n");
+	return (0);
 }
 
 /*
@@ -45,15 +46,14 @@ int	main(int argc, char **argv, char **envp)
 	int		out_fd;
 	int		counter;
 
-	if (argc < 4 || access(argv[1], R_OK) != 0 || \
-					access(argv[argc - 1], W_OK) != 0)
+	if (ft_strncmp(argv[1], "here_doc", 8) == 0)
+		return (here_doc(argc, argv, envp));
+	if (argc < 4 || access(argv[1], R_OK) != 0)
 		exit_error("ERROR: Invalid inputs, check args & file permissions.\n");
-	if (ft_strncmp(argv[1], "here_doc", 8) != 0)
-		here_doc(argc, argv, envp);
 	in_fd = open(argv[1], O_RDONLY);
-	out_fd = open(argv[argc - 1], O_RDWR);
+	out_fd = open(argv[argc - 1], O_RDWR | O_TRUNC | O_CREAT, 0777);
 	if (in_fd < 0 || out_fd < 0)
-		exit_error("ERROR trying to read from input or output files.\n");
+		exit_error("ERROR trying to access input or output files.\n");
 	counter = 1;
 	while (++counter < argc - 2)
 		in_fd = execute_command(in_fd, argv[counter], -1, envp);